Enable to/fromJson to work with non-assignable objects
Previously you could not use toJson() on a polymorphic structure such
as one of those defined in eventcontent.h because it is not assignable
and the default specialisation of JsonObjectConverter used assignment.
To avoid that limitation, one had to specialise JsonObjectConverter for
each descendant of EventContent::Base, which is a lot of boilerplate.

The new JsonConverter (the template underpinning the "default"
to/fromJson implementation) improves on two things:
1. dump() allows to construct your own QJsonObject - or anything else
   convertable to QJsonValue - instead of requiring to fill in the
   pre-constructed one.
2. load() allows to construct your value type directly from QJsonObject
   instead of default-constructing it in advance.