Handle to-device messages before handling roomdata
Probably improves the performance slightly

If we handle to room data first, if a message arrives at the same time as the to-device message containing the key and we handle the message first, it will not be decryptable and stored as undecrypted. Then, when the key is handled, the cache of undecrypted messages is searched, the message decrypted and replaced.

When handling the key first, the message can be decryped instantly.