Drop static factory methods in QOlmSession
Each of them needs QOlmAccount so it's quite natural to have them
as non-static methods of QOlmAccount, as long as QOlmAccount is a friend
class of QOlmSession. In fact, they already are there but used
to delegate the actual work to those static factories; and the callers
mainly already use those QOlmAccount methods.