FixedBuffer<> and PicklingKey
FixedBuffer uses OpenSSL to create a separate heap and store sensitive
material there instead of the common heap. It is not COW and cannot be
reallocated (not for now at least), to ensure that those sensitive
strings don't get copied to other places in the memory.

PicklingKey is a replacement for PicklingMode/Encrypted derived from
`FixedBuffer<char, 128>`.