Apply RAII to PicklingKey and store it in Database
Connection never uses the pickling key directly; and pickling/unpickling
operations always come in conjunction with a respective store/load
operation in Database. So it makes sense to only produce the pickling
key in Connection and then immediately pass it to Database, which is
tasked with pickling/unpickling QOlm* objects in a uniform way
(previously it only did unpickling to load arrays of things like
outbound Megolm sessions).

And since Connection doesn't own the pickling key anymore, it also makes
sense to apply RAII to PicklingKey instead of the two-phase
initialisation used so far.