Move verification functions to QOlmUtility
These never had anything to do with QOlmAccount class, and probably were
only there for convenience of including a single header instead of two.
This is not even the case by now - connection includes both
qolmaccount.h and qolmutility.h anyway. Another change here is getting
rid of free functions, leaving the caller to manage the lifecycle of a
QOlmUtility instance instead of creating and destroying it at every
verification.