Consistently use the QLatin1String key constants
Prepares for the eventual removal of the QString ones.

Deprecation attributes on the QString ones doesn't work unfortunately,
that triggers on every include of event.h (as that is technically a use
then, which is exactly the problem we want to fix by removing them).
This should work though when turning those into extern/exported
declarations.

Another possible approach could be to just make them aliases for the
QLatin1Strings. That wont break ABI as they are not part of the ABI at
this point. It is potentially an API change though, in case of those
keys being used in a place where a QLatin1String would not work. That
does seem somewhat unlikely though, given their primary use is with
QJson*.