WIP Make internal includes work independent of the include search path(s)
This is necessary in order to be able to remove `include/Quotient` from
the include search path to force namespaced includes in consumer code.

This is still vastly incomplete on the non-generated header files, so
consider those only as an example. The real blocker however are the
includes to referenced types in the generated headers.