Make Qt5 and Qt6 versions of libQuotient co-installable
That means renaming the library target and the build system support files
to have a "Qt6" suffix in case of a Qt6 build. Include files remain
unchanged, those work with both versions.

This isn't pretty but should hopefully allow distributions to ship both
variants at the same time for now.