Actually mark global QString constants as deprecated
Technically this is an API change as their type changes from QString to
QLatin1String now, but given that QLatin1String can be used almost
everywhere a QString can be used (and those constants are for use with
the Qt JSON API anyway), this should be fine.

This is not an ABI change as those constants are not part of the ABI but
are copied into every translation unit regardless of their use there.
With that no longer happening now we also save ~12kB of binary size on a
size-optimized build of libQuotient.