run-tests.sh: be dumb and assume autotests/run-tests.sh invocation
BASH_SOURCE, apparently, is not reliable enough to figure out
the script's own directory (in one case CI fails because it's
empty). So just assume that the script is always run from
the directory one level above autotests (which is the way it's
been called in CI for all this time).