Connection::setEncryptionDefault()
This is a way for client authors to govern whether they want
to enable encryption in their Connection objects by default,
instead of calling Connection::enableEncryption() every time
a Connection object is created. This function is intended for
calling very early, in main.cpp or during initial setup of
a client application.

Eventually E2EE will be switched on by default but that's certainly
not for 0.8 (and not for 0.9 maybe even), and then this function
will not be needed anymore.