setup-tests.sh: drop ctest invocation; talk a bit more on exit
The ctest invocation has been a leftover from the
run-tests.sh->setup-tests.sh transformation. The additional
message aims to address a surprise like that of my own when
I forgot that I should set KEEP_SYNAPSE if I don't want the
container to self-destruct on shell exit (that's intentional,
for use in CI).