Fix the logic around FileMetadataMap
When QHash::value() used in FMM::lookup() doesn't find an entry in
the map, it returns a default-constructed EncryptedFileMetadata object
(that is actually invalid), and lookup diligently wraps it in
an Omittable. When the code in MxcReply later checks that Omittable, it
always finds an object inside - and fails to decrypt the file because
that object is invalid. The fixed logic drops Omittable entirely and
officially checks that EncryptedFileMetadata is valid before even trying
to set up for decryption.