Add and document Event::switchOnType()
Before introducing EventMetaType and the respective changes in how
is() works it was necessary to have switchOnType() as a free template
function so that the information on the base event type is not lost
and a fallback visitor could use its fields. Now the entire base event
type hierarchy can be tested with is() and switchOnType() can include,
e.g., overloads accepting StateEvent and RoomEvent one after another;
so it's possible to provide switchOnType() as a member function too.
The tiny difference between calling the member function from the free
one is that because `this` always has type Event now, you can