Deprecate Event::type()
There are too many ways to get various "types" from an event (TypeId,
type(), matrixType(), to name a few). Event::type() should not really
be used in client code: to check the event's actual type in JSON, it's
better to use Event::matrixType(); to match an event against a specific
TypeId, there are is() and switchOnType(); and in rare cases when you
actually need the event type as seen by the metatype system, explicitly
calling Event::metatype() (that Event::type() internally calls) conveys
the intent better.

This commit also fixes the sole case in the library code where type()
was (incorrectly) used.