Split and fix Room::processStateEvent()
This method consists of roughly 3 parts: the preprocessing (that also
checks if further processing should proceed at all - controversially,
see below), the actual change of the event, and the update of internal
structures upon that change. One thing this commit does is splitting out
the first and the third part out, to at least somewhat simplify the
enourmous feed of code (the whole method was 200+ lines long, with
plenty of branching inside). Another thing is fixing the problem in
preprocessing; the fact that it's conditional and can therefore
prevent the actual replacement of the event comes at odds with its usage
from the event redaction logic - leading to #665. The condionality
was introduced to avoid breaking the internal structures with invalid
event content coming from the wire; but the case that leads to #665 has
nothing to do with internal structures protection, it was just an
optimisation on an event that is no-op from those internal structures
standpoint (no changes in display name/avatar). No-op for them doesn't
mean the event shouldn't be refreshed though, and to stay in sync with
the server it's actually better to insert the new event instead of
keeping the old one. The same problem exists in account data code but
we don't have redactions there (yet, at least) so it was kind of fine.
It will still be addressed in a separate commit, with the same
rationale, being better aligned with the homeserver.

Fixes #665.