asKeyValueRange: support accepting rvalues
Since it is intended to be used with Qt containers which are
implicitly-shared, asKeyValueRange can copy-construct an internal
snapshot for cheap, for the sake of lifetime extension. However,
any changes on that copy would not propagate to the original container
rvalue. To avoid surprises, non-const begin/end adaptors are only
enabled with (non-const) lvalues.