Split out logging categories definitions
The header logging.h mixed client-facing and library-internal things.
This commit puts library-internal things (which happened to be
definitions of logging categories and nothing else) into
logging_categories_p.h and client-facing things landed in util.h as
there not many of them (a couple of debug manipulators and operator<<
overloads).