CMakeLists + gtad.yaml: Enable key_backup.*
key_backup.yaml has a nasty repetition of the 'version' parameter within
a single API call, with the comment saying that if the version replica
in the request body is specified it should coincide with the version
provided in the request path. By default, GTAD would diligently pull
both replicas to the constructor argument list, producing incorrect
C++ code. The line added to gtad.yaml here eliminates the replica
in the request body (because '>' only applies to the request body -
GTAD documentation covers that, although not quite explicitly).