cryptoutils: check sizes; constify where possible
checkedSize() makes sure that the passed 64-bit sizes can fit signed int
parameters commonly accepted throughout OpenSSL. When the size ends up
clamped, the crypto function that uses it will likely either fail (e.g.
in case of decryption) or act on the trimmed payload (encryption,
digests) but that will be the problem the caller will have to deal with.