Fix a regression in handling attachments
aesCtr256En/Decrypt used FixedBuffer for the target (encrypted and
decrypted, respectively) buffer. FixedBuffer uses secure memory, and
it is purposefully limited to 64k, because it's meant to be used for
secrets, not for payloads (for now at least). This commit reverts to
using normal QByteArray to store the result, using asWritableCBytes()
to get a span adapter that is more convenient to work with in OpenSSL.