Refactor unlockAndLoad()
The method became rather large and with additional error checking would
get even larger - so now the initial preparations concerned with
default key and key description retrieval are put out to
UnlockData::prepare() - a static factory method returning the loaded
information that is used further in the process. That allows to put
the piece concerned with making the key out of the passphrase to
unlockSSSSWithPassphrase() (formerly not quite correctly named as
unlockSSSSFromPassword). More byte_views are introduced, to ensure
correct payload sizes as early (preferably at compile time) as possible.