Make sure Quotest doesn't get stuck on exit
There was a wrong assumption that the message with the final report
will always get sent last. This is not true when 429's occur, as
the library may reorder sending events (related issue: #133). Moreover,
this actually was premature optimisation because the code checks for
messages reaching the homeserver anyway. Once these two lines are gone,
the code starts doing exactly what the comments next to it say, and
exits promptly after the last message reaches the homeserver.