Simplify connectUntil()
Now that connectSingleShot() is implemented in Qt we don't need extra
complexity consumed by _impl::connect(), we can just build a lambda
inside connectUntil() and use it as a slot. Connection lifecycle is
now controlled by a context QObject instead of putting a Connection
object on the heap - the code is simpler and the performance
implications are negligible. Also, the slot argument is now forwarded
rather than copied, to support uncopyable function objects.

Oh, and the PmfSlot has been out of use since we got rid of Qt 5.