JobHandle
Because this commit changes the returned type of Connection::callApi()
to JobHandle<JobT>, it will break the code that initialises an `auto*`
variable from Connection::callApi(). If the variable has a type of
`SpecificJob*`, implicit conversion of QPointer should maintain
compatibility; and if it's `auto` it can be later used as an ordinary
QPointer, too. But now you get the QFuture interface at `job.`, while
still having the job interface at `job->`.