Add no_unique_address where appropriate/works
Skip is an empty structure, so we can optimise away pretty much
the entire BoundFn structure when it's initialised with {Skip, Skip}
(a case when then() is called with the single onSuccess argument).

Unfortunately, Clang either crashes with ICE or emits faulty code when
fn has no unique address (maybe because BoundFn::call() takes its
address to determine if it's a function pointer? I don't know); so that
no_unique_address is only added for compilers that can run with it.