Store senderkey for inbound megolm sessions
I don't know of any case where we actually need those locally, but for creating correct key backups (or exports), we need to fill it in.
For sessions we have received locally from olm, we can reconstruct it (as this patch does during database migration), but for key backup,
we have to store it.