findIndirect()
I was surprised not to find some kind of dereferencing filter in
std::ranges, and since making custom views still doesn't quite work
as of C++23-tinted C++20 that we have for now, I made an algorithm
instead.

Curiously, GNU libstdc++ has a bug/feature where std::find[_if]()
dereferences the element in the range for your projection if (and only
if!) you use that projection is a member function pointer - acting as
findIndirect() in that particular situation.