ALARM() and ALARM_X() facilities
I always had a mental struggle with assertion functions that print
a message: the assertion fails when the condition is false but
the message is also printed when the condition fails. Moreover,
for some reason runtime assert functions have no way of tackling stuff
in Release builds: if you want to somehow bail out in a Release build,
you have to separately check the condition and then put Q_ASSERT(false)
or its equivalent. So here's a "negative assertion" facility - it does
have footprint even in Release builds but also helps to somehow move on,
rather than just let the Release build crash on an eventual null pointer
dereferencing.