Room: postEvent() is deprecated, long live post()
Technically very similar to postEvent(), post() takes ownership
guarantees to the type system, and returns a reference to the pending
event item that is produced in the process of posting anyway and has
all the information in a single structure including the transaction id
that is returned now - instead of the sole transaction id which is
frequently only used to find the pending event item first thing.

The template overload of post() also allows to construct an event
on the fly instead of creating an event object upfront (see
the corresponding change in quotest.cpp).

Other post*() methods still return transaction id in order to retain
interoperability with QML. I haven't seen Quaternion or NeoChat actually
using the returned values from post*() calls though. We might
eventually move on to some Q_GADGET wrapper of PendingEventItem so that
post*() methods could still be fully interoperable with QML but provide
more information that a mere transaction id - if needed.