BaseJob: rebalance back-off timeouts
Fail with the first request sooner, wait on the last one longer. Driven
by Synapse getting stuck with an initial login flows request for some
reason; the next request goes through immediately though, so this
rebalancing helps to speed up things a bit.

Correspondingly, the timeout in testutils.cpp (do we even need it?) is
increased to give time for the second attempt to take place.