Remove needsToken logic from ConnectionData
This was introduced as a workaround for cases when homeservers
require authentication on media endpoints, deviating from the spec text
of that time in order to protect themselves. This workaround is not
necessary since #776; DownloadFileJob and MediaThumbnailJob now always
supply the access token even when the homeserver doesn't require it.