Fix printable(); it's now QUO_CSTR() and a macro
The first take on a "better version of q(Utf8)Printable" was actually
worse as it returned a pointer to a local temporary, and it was mere
luck that it never backfired (or rather, we don't know it backfired -
all its usages are in failure paths, after all). The rework reverts to
using a macro on the top level, using two helpers to first make that
temporary buffer, this time on the macro level, and then extract
a pointer to the data in that buffer. The rest of the commit streamlines
usage of this macro (and switches to it from qPrintable in a couple of
places).