Fix GCC -Wattributes warnings
GCC gets concerned that JobHandle<>::BoundFn is not exported by the
library although lambdas it carries are (because they are created in
Connection that is tagged with QUOTIENT_API). JobHandle is an inline
class template but it's not too bad (and from what I gather, _is_
correct) to add QUOTIENT_API on it, too.