Quotient::isStateEvent(); move Event::isStateEvent() to RoomEvent
Checking whether an event is a state event only makes sense in room
context, hence the moving. As for Quotient::isStateEvent(), this checks
an event type id against known event types that are derived from
StateEvent; the function will be used to simplify the API to check power
levels in a later commit.