TestOlmAccount: rewrite with futures; ease timeouts
In some situations Synapse takes a lot of time and sometimes even fails
to respond to a request at once; to alleviate that, most timeouts are
increased from 10 to 40 seconds, to give time for a retry. And to unify
the waiting code, Quotient::waitForFuture() is introduced that
internally uses QTest::qWaitFor() that is more lightweight than
QSignalSpy.