Rework RoomMessageEvent content API again
It's really straightforward now :) RoomMessageEvent::has<>() - check
that it's possible to extract event content of a given type;
RoomMessageEvent::get<>() - actually try to get event content of that
type. At this point has() is not optimised and in some cases effectively
creates a content object any way. Can be tackled sometime later.